requirements.txt
Flask
python-dotenv
requests
openai>=1.0
psycopg2-binary
